iostream -> stdio.h
The C++ "stream" functions in the code have been changed with their C
counterparts (e.g. instead of writing to std::cerr now I use fputs and
fprintf). This allows a much easier formatting of output file names, and
it is needed for implementing further changes to the code I have in
mind.